pyproject.toml
[project]
name = "shadowfax-flash-mcp-server"
version = "0.1.0"
description = "It is a MCP server for Shadowfax Flash"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[client]>=1.12.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
